Refactor with renaming
