+++ src/main/resources/view/MainWindow.fxml
@@ -11,7 +11,7 @@
             prefHeight="600.0" prefWidth="400.0"
             xmlns="http://javafx.com/javafx/17"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="Bunbun.Gui.MainWindow">
+            fx:controller="bunbun.gui.MainWindow">
     <children>
         <TextField fx:id="userInput"
                    layoutY="558.0"
